# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateSkillIn": {
        "additionalProperties": false,
        "properties": {
          "files": {
            "additionalProperties": {
              "type": "string"
            },
            "maxProperties": 200,
            "minProperties": 1,
            "title": "Files",
            "type": "object"
          },
          "publish": {
            "default": false,
            "title": "Publish",
            "type": "boolean"
          },
          "reason": {
            "default": "",
            "maxLength": 2000,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "files"
        ],
        "title": "CreateSkillIn",
        "type": "object"
      },
...
      "SkillLibraryErrorOut": {
        "description": "Every refusal. `issues` accompanies `invalid_bundle`; `reasons` accompanies `publish_blocked`.",
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "issues": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BundleIssueOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issues"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reasons": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasons"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "SkillLibraryErrorOut",
        "type": "object"
      },
      "SkillListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SkillSummaryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "SkillListOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Felix",
      "url": "https://docs.felix.run/"
    },
    "description": "Felix compiles an `apiVersion: felix/v1` manifest into a governed agent with durable fibers, memory, skills, eval, approvals, and sandboxes \u2014 exposed over OpenAI-compatible, A2A, MCP, and REST/SSE surfaces. Fork, rewind, and steer live runs.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://github.com/felix-run/felix/blob/main/LICENSE"
    },
    "summary": "Self-hostable managed agents harness.",
    "title": "Felix"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/skill-library": {
      "get": {
        "description": "The tenant's library skills by name, each with its live and newest version.\n\nPages by name. The filters apply to a page after it is read, so a filtered page can be\nshort; keep following `next_cursor` until it is null.",
        "operationId": "list_library_skill_library_get",
        "parameters": [
          {
            "description": "`live`: has a live version. `draft`: has drafts awaiting review. `archived`: neither.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "live",
                    "draft",
                    "archived"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`live`: has a live version. `draft`: has drafts awaiting review. `archived`: neither.",
              "title": "Status"
            }
          },
          {
            "description": "Who wrote the newest version.",
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "agent",
                    "operator"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Who wrote the newest version.",
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Library",
        "tags": [
          "Skill library"
        ]
      },
      "post": {
        "description": "Save a new skill as an operator draft (version 0.1.0), and publish it if `publish`.\n\nThe name is the SKILL.md's own. 409 `skill_exists` if the library already holds it --\n`PUT /{name}/versions` adds a version.",
        "operationId": "create_library_skill_skill_library_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkillIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillWriteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create Library Skill",
        "tags": [
          "Skill library"
        ]
      }
    },
    "/skill-library/policy": {
      "get": {
        "description": "The gate every publish and rollback passes. From settings for now\n(`FELIX_SKILL_PUBLISH_MIN_QUALITY`, `FELIX_SKILL_PUBLISH_BLOCK_ON_ADVISORY`).",
        "operationId": "publish_policy_skill_library_policy_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillPolicyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Publish Policy",
        "tags": [
          "Skill library"
        ]
      }
    },
    "/skill-library/review": {
      "get": {
        "description": "Every draft in the tenant awaiting a decision, oldest first, across skills.",
        "operationId": "review_queue_skill_library_review_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewQueueOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillLibraryErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Review Queue",
        "tags": [
          "Skill library"
        ]
      }
    },
...
  }
}